\name{dichotomous}
\alias{dichotomous}
\docType{data}
\title{
Raw data from a multiple-choice test
}
\description{
The data is taken from the following website http://www.jmetrik.com/example-data.php.
This file includes nominal responses of 6,000 examinees to 56 binary items.
}
\usage{data(dichotomous)}
\format{
  A data frame with 60000 examinees and 56 items
}
