\name{timms2011_usa}
\alias{timss2011_usa}
\docType{data}
\title{
TIMMS 2011 USA Data - Attitude Towards Math
}
\description{
The data is a subset of TIMSS 2011 USA data and includes responses for six statements to measure attitudes towards math.
These rating scale items have response codes from 0 to 3 with 0 indicating "I strongly disagree", 1 indicating "I disagree", 2 indicating "I agree", and 3 indicating "I strongly agree" for a given statement. Note that items 14B and 14C has to be reverse coded before analysis to make them consistent with other four items.
}
\usage{data(timss2011_usa)}
\format{
  A data frame with 10079 observaiton and 6 items.
  \describe{
    \item{\code{Q14A}}{I enjoy learning mathematics}
    \item{\code{Q14B}}{I wish have not to study Math}
    \item{\code{Q14C}}{Mathematics is boring}
    \item{\code{Q14D}}{I learn interesting things in mathematics class}
    \item{\code{Q14E}}{I like mathematics}
    \item{\code{Q14F}}{I think it's important to do well in mathematics}
  }
}
