\name{itemanalysis}
\alias{itemanalysis}
\alias{itemanalysis}
\docType{package}
\title{
itemanalysis: Classical Test Theory Item Analysis
}
\description{
This package
}
\details{
\tabular{ll}{
Package: \tab itemanalysis\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2022-06-13\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
The package can be used to run classical item analysis for multiple-choice test items and polytomously scored items (e.g., rating scale items).
}
\author{
Maintainer: Cengiz Zopluoglu <cen.zop@gmail.com>
}
\keyword{ package }
\seealso{
\code{\link{itemanalysis1}} for running classical item analysis for multiple-choice test items
\code{\link{itemanalysis2}} for running classical item analysis for polytomously scored items
}